Development

+++ Project/Logic/RoomsLogic.cs
@@ -2,7 +2,7 @@ public static class RoomsLogic
 {
     private static RoomsAccess _access = new();
 
-    public static RoomModel GetRoomById(int id)
+    public static RoomModel GetRoomById(Int64 id)
     {
         return _access.GetById(id);
     }
